Update call params JSdoc
