+++ src/internal/compilerSupport/v1.ts
@@ -2,14 +2,4 @@ export { FluencePeer } from '../FluencePeer';
 export { ResultCodes } from '../../internal/CallServiceHandler';
 export { RequestFlow } from '../../internal/RequestFlow';
 export { RequestFlowBuilder } from '../../internal/RequestFlowBuilder';
-import { SecurityTetraplet } from '@fluencelabs/avm';
-import { PeerIdB58 } from '../commonTypes';
-
-export interface CallParams<ArgName extends string | null> {
-    particleId: string;
-    initPeerId: PeerIdB58;
-    timeStamp: number;
-    ttl: number;
-    signature: string;
-    tetraplets: { [key in ArgName]: SecurityTetraplet[] };
-}
+export { CallParams } from '../commonTypes';
